removed releases and contributors since they cannot be retrieved
Contributors are not loaded immediately. At first "fetching contributors"
is showed instead.
Releases cannot be retrieved too because sometimes contributors are loaded
and sometimes they are not and so the scraper cannot understand if a projects
tab is present or not.
Since releases are showed next to projects tab you cannot know if they are
releases or contributors